fix: Use uint32_t instead of size_t for SDCC compatibility
Replace `size_t *count` parameter in stc89c52rc_test_registry() with `uint32_t *count`
for better SDCC compiler compatibility. SDCC has issues with size_t in certain
contexts, particularly with pointer-to-pointer patterns.

Updated:
  - include/stc89c52rc/cli/test_registry.h: Function declaration
  - src/cli/test_registry.c: Function implementation
  - src/cli/cli_commands.c: Function call and loop variable

This fixes the syntax error in test_registry.h:27 when compiling with SDCC.

+++ include/stc89c52rc/cli/test_registry.h
@@ -24,7 +24,7 @@ typedef struct {
  * @param count Output: number of tests in registry
  * @return Pointer to array of test cases
  */
-const stc89c52rc_test_case_t *stc89c52rc_test_registry(size_t *count);
+const stc89c52rc_test_case_t *stc89c52rc_test_registry(uint32_t *count);
 
 /**
  * Find a test by name (case-sensitive)
